lm-test/pyproject.toml
[tool.poetry]
name = "lm-test"
version = "0.1.0"
description = "License Manager automated integration test"
authors = ["Omnivector Solutions <info@omnivector.solutions>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/omnivector-solutions/license-manager"
packages = [{include = "lm_test"}]

[tool.poetry.dependencies]
python = "^3.11"
license-manager-cli = {path = "../lm-cli"}
py-buzz = "^3.2.1"
httpx = "^0.22.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.961"
pyproject-flake8 = "^0.0.1-alpha.4"

[tool.poetry.scripts]
test-lm = "lm_test.main:main"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
lines_after_imports = "2"
profile = "black"
multi_line_output = 3
include_trailing_comma = true

[tool.flake8]
max-line-length = 120
max-complexity = 40

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
